+++ src/RecoveryServices/RecoveryServices.Backup.Test/ScenarioTests/IaasVm/PolicyTests.ps1
@@ -15,6 +15,7 @@
 $resourceGroupName = "RecoveryServicesBackupTestRg";
 $resourceName = "PsTestRsVault";
 $policyName = "PsTestPolicy";
+$defaultPolicyName = "DefaultPolicy";
 $DefaultSnapshotDays = 2;
 $UpdatedSnapShotDays = 5;
 
@@ -52,6 +53,16 @@ function Test-AzureVMPolicy
 			-Name $policyName
 		Assert-NotNull $policy
 		Assert-AreEqual $policy.Name $policyName
+
+		$defaultPolicy = Get-AzRecoveryServicesBackupProtectionPolicy `
+			-VaultId $vault.ID `
+			-Name $defaultPolicyName
+		Assert-NotNull $defaultPolicy
+		Assert-AreEqual $defaultPolicy.Name $defaultPolicyName
+		Assert-True { $defaultPolicy.SchedulePolicy.ScheduleRunDays -contains "Saturday" }
+		Assert-True { $defaultPolicy.SchedulePolicy.ScheduleRunDays -contains "Thursday" }
+		Assert-False { $defaultPolicy.SchedulePolicy.ScheduleRunDays -contains "Sunday" }
+		Assert-False { $defaultPolicy.SchedulePolicy.ScheduleRunDays -contains "Friday" }
 
 		# Get default policy objects (this data is generated partially at random. So, running this again gives different values)
 		$schedulePolicy = Get-AzRecoveryServicesBackupSchedulePolicyObject -WorkloadType AzureVM